# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api eccomerce",
        "contact": {
            "email": "guilhermeieski@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/blogs/{empresa_id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Listar blogs por empresa",
                "description": "Retorna uma lista paginada de blogs pertencentes a uma empresa específica, com a opção de aplicar filtros usando query parameters.",
                "operationId": "6efe1affe67f396771bb45bd88067310",
                "parameters": [
                    {
                        "name": "empresa_id",
                        "in": "path",
                        "description": "ID da empresa cujos blogs serão listados",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca para filtrar blogs",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de blogs",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar blogs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar blogs"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/blog/criar": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Criar um novo blog",
                "description": "Cria um novo blog com os dados fornecidos e faz o upload da imagem associada, se presente.",
                "operationId": "964a0787bb909fa02b74d2f47e3a2aae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "texto"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Título do Blog"
                                    },
                                    "texto": {
                                        "type": "string",
                                        "example": "Conteúdo do blog."
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "byte",
                                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Blog criado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar blog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao criar blog"
                                        },
                                        "erro": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/blog/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Exibir detalhes de um blog",
                "description": "Retorna os detalhes de um blog específico, incluindo a URL da foto armazenada no S3.",
                "operationId": "6ac9d2eb1f1ccbc64de4ae3a2bb02941",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do blog a ser exibido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do blog",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Blog não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Blog não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar blog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar blog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/blog/editar/{id}": {
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Atualizar um blog",
                "description": "Atualiza os dados de um blog existente, incluindo a imagem associada.",
                "operationId": "bdc7e678d9f34194f9291bc387464b82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do blog a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "texto"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Novo Título do Blog"
                                    },
                                    "texto": {
                                        "type": "string",
                                        "example": "Conteúdo atualizado do blog."
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "byte",
                                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Blog atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Blog não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Blog não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar blog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao atualizar blog"
                                        },
                                        "erro": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/blog/delete/{id}": {
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Deletar um blog",
                "description": "Remove um blog específico, incluindo a exclusão da imagem associada no S3, caso exista.",
                "operationId": "1b2e400ab40bc55cfaf59cc369b5227b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do blog a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Blog deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Blog não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Blog não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar blog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao deletar blog"
                                        },
                                        "erro": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}